requirements.txt
asgiref==3.2.5
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
decorator==4.4.2
Django==3.0.4
django-location-field==2.1.0
future==0.18.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-telegram-bot==12.4.2
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
zope.interface==5.0.0
